% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread.heckmange.R
\name{bread.heckmange}
\alias{bread.heckmange}
\title{bread.heckmange}
\usage{
bread.heckmange(object, ...)
}
\arguments{
\item{object}{An object of class \code{fitheckmange}, which is the result of fitting a Heckman selection model using the \code{fitheckmange} function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A matrix representing the bread component of the sandwich estimator. The matrix is calculated as the product of the number of observations and the variance-covariance matrix of the estimated parameters.
}
\description{
Bread Function for the \code{fitheckmange} Model
}
\details{
This function calculates the "bread" component of the sandwich estimator for the \code{fitheckmange} model. The bread matrix is typically defined as the product of the number of observations and the variance-covariance matrix of the estimated parameters.

The bread matrix is an essential component of the sandwich estimator used to obtain robust standard errors. It reflects the variability in the estimated parameters due to the model's fit. The function uses the number of observations and the variance-covariance matrix from the \code{fitheckmange} model object to compute this matrix.
}
