% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.fastHeckmanGe.R
\name{coef.fastHeckmanGe}
\alias{coef.fastHeckmanGe}
\title{coef.fastHeckmanGe}
\usage{
\method{coef}{fastHeckmanGe}(object, part = c("selection", "outcome", "dispersion", "correlation"))
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{part}{A charecter vector with the model part names: selection, outcome, dispersion, correlation. By default it returns the complete coefficient vector, for all parts.}
}
\value{
Coefficients extracted from the model object.
}
\description{
Extract Generalized Heckman Model Coefficients
}
