% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastHeckmanGe.R
\name{fastHeckmanGe}
\alias{fastHeckmanGe}
\title{fastHeckmanGe}
\usage{
fastHeckmanGe(
  selection,
  outcome,
  dispersion,
  correlation,
  data = sys.frame(sys.parent()),
  weights = NULL,
  cluster = NULL,
  start = NULL
)
}
\arguments{
\item{selection}{A formula. Selection equation.}

\item{outcome}{A formula. Outcome Equation.}

\item{dispersion}{A right-handed formula. The equation for fitting of the Dispersion Parameter.}

\item{correlation}{A right-handed formula. The equation for fitting of the Correlation Parameter.}

\item{data}{A data.frame.}

\item{weights}{an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.}

\item{cluster}{a variable indicating the clustering of observations, a list (or data.frame) thereof, or a formula specifying which variables from the fitted model should be used. See documentation for sandwich::vcovCL}

\item{start}{Optional. A numeric vector with the initial values for the parameters.}
}
\value{
#' Returns... several things... # TEMOS QUE DESCREVER AQUI
}
\description{
Estimates the parameters of the Generalized Heckman model
}
\details{
Optimized Function for fitting the Generalized Heckman Model

(original version: package ssmodels. Modified by Rogerio Barbosa)

The fastHeckmanGe() function fits a generalization of the Heckman sample
selection model, allowing sample selection bias and dispersion parameters
to depend on covariates.
}
\examples{
data(MEPS2001)
selectEq  <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
dispersion  <- ~ age + female + totchr + ins
correlation  <- ~ age
system.time({fit.fast = fastHeckmanGe(selectEq, outcomeEq, dispersion, correlation, data = MEPS2001)})

}
