% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfastHeckmanGe.R
\name{fitfastHeckmanGe}
\alias{fitfastHeckmanGe}
\title{fitfastHeckmanGe}
\usage{
fitfastHeckmanGe(start, YS, XS, YO, XO, Msigma, Mrho, w)
}
\description{
Optimized Function for fitting the Generalized Heckman Model
}
\details{
(original version: package ssmodels. Modified by Rogerio Barbosa)
}
