% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.fastHeckmanGe.R
\name{fitted.fastHeckmanGe}
\alias{fitted.fastHeckmanGe}
\title{fitted.fastHeckmanGe}
\usage{
\method{fitted}{fastHeckmanGe}(object, part = c("selection", "outcome", "dispersion", "correlation"))
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{part}{A character vector with the model part names: selection, outcome (default), dispersion, correlation.}
}
\value{
Fitted values extracted from the fastHeckmanGe object.
}
\description{
Extract Fitted Values of the Generalized Heckman Model
}
