% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.heckmanGE.R
\name{fitted.heckmanGE}
\alias{fitted.heckmanGE}
\title{fitted.heckmanGE
Extract Fitted Values of the Generalized Heckman Model}
\usage{
\method{fitted}{heckmanGE}(
  object,
  part = c("selection", "outcome", "dispersion", "correlation"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{heckmanGE}. This object should contain the fitted values for different parts of the model.}

\item{part}{A character vector specifying which part of the model to return the fitted values for. Options are "selection", "outcome" (default), "dispersion", or "correlation". If multiple parts are provided, only the "outcome" part will be returned.}

\item{\dots}{Additional arguments passed to or from other methods. These are not used in this method but must be included to match the generic method signature.}
}
\value{
A vector of fitted values corresponding to the specified part of the \code{heckmanGE} model. The type of the returned values depends on the part specified.
}
\description{
This function extracts the fitted values from a \code{heckmanGE} object. You can specify which part of the model you want to retrieve the fitted values for: selection, outcome, dispersion, or correlation. By default, it returns the fitted values for the outcome part of the model.
}
\details{
\itemize{
\item If \code{part} is "selection", the function returns the fitted values from the selection part of the model.
\item If \code{part} is "outcome", the function returns the fitted values from the outcome part of the model.
\item If \code{part} is "dispersion", the function returns the fitted values from the dispersion part of the model.
\item If \code{part} is "correlation", the function returns the fitted values from the correlation part of the model.
}

If \code{part} is not one of the specified options, an error will be raised. If multiple parts are provided, the function defaults to returning the fitted values for the outcome part of the model.
}
