% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/heckmange.R
\name{heckmange}
\alias{heckmange}
\title{heckmange: A Package for Fitting Sample Selection Models}
\usage{
heckmange(
  selection,
  outcome,
  dispersion,
  correlation,
  data = sys.frame(sys.parent()),
  weights = NULL,
  cluster = NULL,
  start = NULL
)
}
\arguments{
\item{selection}{A formula. Selection equation.}

\item{outcome}{A formula. Outcome Equation.}

\item{dispersion}{A right-handed formula. The equation for fitting of the Dispersion Parameter.}

\item{correlation}{A right-handed formula. The equation for fitting of the Correlation Parameter.}

\item{data}{A data.frame.}

\item{weights}{an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.}

\item{cluster}{a variable indicating the clustering of observations, a list (or data.frame) thereof, or a formula specifying which variables from the fitted model should be used. See documentation for sandwich::vcovCL}

\item{start}{Optional. A numeric vector with the initial values for the parameters.}
}
\value{
A list of results from the fitted model, including parameter estimates, the Hessian matrix, number of observations, and other relevant statistics. If initial values are not provided, the function estimates them using the Heckman two-step method.

#' Returns... several things... # TEMOS QUE DESCREVER AQUI
}
\description{
This package provides functions for fitting sample selection models, specifically the Heckman-Ge model. It includes functionality for specifying selection and outcome equations, as well as adjusting parameters for dispersion and correlation.

Estimates the parameters of the Generalized Heckman model
}
\details{
Optimized Function for fitting the Generalized Heckman Model

(original version: package ssmodels. Modified by Rogerio Barbosa)

The heckmange() function fits a generalization of the Heckman sample
selection model, allowing sample selection bias and dispersion parameters
to depend on covariates.
}
\examples{
data(MEPS2001)
selectEq  <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
dispersion  <- ~ age + female + totchr + ins
correlation  <- ~ age
}
\author{
Fernando de Souza Bastos
}
\keyword{Heckman,}
\keyword{Sample}
\keyword{Selection}
