% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meatCL.fastHeckmanGe.R
\name{meatCL.fastHeckmanGe}
\alias{meatCL.fastHeckmanGe}
\title{meatCL.fastHeckmanGe}
\usage{
meatCL.fastHeckmanGe(
  x,
  cluster = NULL,
  type = NULL,
  cadjust = TRUE,
  multi0 = FALSE,
  ...
)
}
\description{
meatCL.fastHeckmanGe
}
