% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.fastHeckmanGe.R
\name{model.frame.fastHeckmanGe}
\alias{model.frame.fastHeckmanGe}
\title{model.frame.fastHeckmanGe}
\usage{
\method{model.frame}{fastHeckmanGe}(
  object,
  part = c("selection", "outcome", "dispersion", "correlation")
)
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{part}{A character vector with the model part names: selection, outcome (default), dispersion, correlation.}
}
\value{
A model frame of (part of) a fastHeckmanGe object.
}
\description{
Get the Model Frames of a Generalized Heckman Regression
}
