% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.fastHeckmanGe.R
\name{model.matrix.fastHeckmanGe}
\alias{model.matrix.fastHeckmanGe}
\title{model.matrix.fastHeckmanGe}
\usage{
\method{model.matrix}{fastHeckmanGe}(
  object,
  part = c("selection", "outcome", "dispersion", "correlation")
)
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{part}{A character vector with the model part names: selection, outcome (default), dispersion, correlation.}
}
\value{
A model matrix of (part of) a fastHeckmanGe object.
}
\description{
Get the Design Matrices of a Generalized Heckman Regression
}
