% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fastHeckmanGe.R
\name{predict.fastHeckmanGe}
\alias{predict.fastHeckmanGe}
\title{predict.fastHeckmanGe}
\usage{
\method{predict}{fastHeckmanGe}(
  object,
  part = c("selection", "outcome", "dispersion", "correlation"),
  newdata = NULL,
  type = c("link", "response"),
  cofint = F,
  confidence_level = 0.95
)
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{part}{A character vector with the model part names: selection, outcome (default), dispersion, correlation.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{type}{the type of prediction required. The default (link) is on the scale of the linear predictors; the alternative "response" is on the scale of the response variable after applying the inverse link function .}

\item{cofint}{A logical. If true, the function returns a matrix, with the mean predicted value, and the lower and upper bounds of the confidence interval.}

\item{confidence_level}{A numeric value. The confidence level for the Confidence Interval of the prediction (if cofint is set to TRUE). Default: 0.95.}
}
\value{
Predictions from fastHeckmanGe the object.
}
\description{
Predictions from the Generalized Heckman Model
}
