% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fastHeckmanGe.R
\name{print.fastHeckmanGe}
\alias{print.fastHeckmanGe}
\title{print.fastHeckmanGe}
\usage{
\method{print}{fastHeckmanGe}(object, ...)
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{...}{others functions.}
}
\value{
Print estimates of the Generalized Heckman model
}
\description{
Print of Generalized Heckman Model results
}
