% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.heckmange.R
\name{print.heckmange}
\alias{print.heckmange}
\title{print.heckmange
Print of Generalized Heckman Model Results}
\usage{
\method{print}{heckmange}(x, ...)
}
\arguments{
\item{x}{An x of class \code{heckmange}. This x should be a fitted model whose results are to be printed.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Prints the estimates and statistics of the Generalized Heckman model to the console.
}
\description{
Prints a summary of the results from a fitted \code{heckmange} model, including estimates for different model components, log-likelihood, AIC, BIC, and other relevant statistics.
}
