% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.fastHeckmanGe.R
\name{residuals.fastHeckmanGe}
\alias{residuals.fastHeckmanGe}
\title{residuals.fastHeckmanGe}
\usage{
\method{residuals}{fastHeckmanGe}(object, part = c("selection", "outcome"))
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{part}{A charecter vector with the model part names: either 'selection' or 'outcome' (default).}
}
\value{
Residuals extracted from the object object.
}
\description{
Extract Residuals of the Generalized Heckman Model
}
