% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.heckmanGE.R
\name{residuals.heckmanGE}
\alias{residuals.heckmanGE}
\title{residuals.heckmanGE
Extract Residuals of the Generalized Heckman Model}
\usage{
\method{residuals}{heckmanGE}(object, part = c("selection", "outcome"), ...)
}
\arguments{
\item{object}{An object of class \code{heckmanGE}. This object should be a fitted model from which residuals are to be extracted.}

\item{part}{A character vector specifying which model component's residuals to return: either 'selection' or 'outcome'. Defaults to 'outcome'.}

\item{\dots}{Additional arguments passed to or from other methods. These are not used in this method but must be included to match the generic method signature.}
}
\value{
A vector of residuals extracted from the specified part of the \code{heckmanGE} model.
}
\description{
Extracts residuals from a fitted \code{heckmanGE} model for a specified model component.
}
