% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.fastHeckmanGe.R
\name{sandwich.fastHeckmanGe}
\alias{sandwich.fastHeckmanGe}
\title{sandwich.fastHeckmanGe}
\usage{
sandwich.fastHeckmanGe(
  x,
  bread. = bread.fastHeckmanGe,
  meat. = meat.fastHeckmanGe,
  ...
)
}
\description{
sandwich.fastHeckmanGe
}
