% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.heckmanGE.R
\name{sandwich.heckmanGE}
\alias{sandwich.heckmanGE}
\title{sandwich.heckmanGE
Sandwich Estimator for Generalized Heckman Model}
\usage{
sandwich.heckmanGE(x, bread. = bread.heckmanGE, meat. = meat.heckmanGE, ...)
}
\arguments{
\item{x}{An object of class \code{heckmanGE} representing a fitted Generalized Heckman model.}

\item{bread.}{A function to compute the "bread" part of the sandwich estimator. Defaults to \code{bread.heckmanGE}.}

\item{meat.}{A function to compute the "meat" part of the sandwich estimator. Defaults to \code{meat.heckmanGE}.}

\item{...}{Additional arguments passed to \code{meat.}.}
}
\value{
A variance-covariance matrix for the \code{heckmanGE} model, computed using the sandwich estimator.
}
\description{
Computes the sandwich estimator of the variance-covariance matrix for the \code{heckmanGE} model.
}
