% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.R
\name{step2}
\alias{step2}
\title{step2: Two-Step Estimation Function}
\usage{
step2(YS, XS, YO, XO, Msigma, Mrho, w)
}
\arguments{
\item{YS}{A binary vector indicating selection (1 if selected, 0 otherwise).}

\item{XS}{A matrix of covariates for the selection equation.}

\item{YO}{A vector of observed outcomes for the selected sample.}

\item{XO}{A matrix of covariates for the outcome equation.}

\item{Msigma}{A matrix of covariates for the dispersion equation.}

\item{Mrho}{A matrix of covariates for the correlation structure.}

\item{w}{A vector of weights to be used in the estimation process.}
}
\value{
A list containing the estimated coefficients for the selection, outcome, dispersion, and correlation equations.
}
\description{
This function performs a two-step estimation process, commonly used in models that require correction for sample selection bias. It estimates parameters for selection, outcome, and dispersion equations.
}
\details{
The function first estimates the selection equation using a probit model. The inverse Mills ratio (IMR) is calculated and included as an additional covariate in the outcome and dispersion equations. The outcome equation is estimated using weighted least squares, and the residuals are used to estimate the dispersion equation. Initial guesses for the correlation structure are also provided.
}
