% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fastHeckmanGe.R
\name{summary.fastHeckmanGe}
\alias{summary.fastHeckmanGe}
\title{summary.fastHeckmanGe}
\usage{
\method{summary}{fastHeckmanGe}(object, ...)
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{...}{others functions.}
}
\value{
Print estimates of the parameters of the Optimized Generalized Heckman model
}
\description{
Summary of Generalized Heckman Model
}
