% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.heckmanGE.R
\name{summary.heckmanGE}
\alias{summary.heckmanGE}
\title{summary.heckmanGE
Summary of Generalized Heckman Model}
\usage{
\method{summary}{heckmanGE}(object, ...)
}
\arguments{
\item{object}{An object of class \code{heckmanGE} representing a fitted Generalized Heckman model.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
Prints a detailed summary of the fitted Generalized Heckman model, including parameter estimates, standard errors, model fit statistics, and optimization details.
}
\description{
Provides a summary of the parameters and diagnostic information from a fitted Generalized Heckman model.
}
