% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.fastHeckmanGe.R
\name{vcov.fastHeckmanGe}
\alias{vcov.fastHeckmanGe}
\title{vcov.fastHeckmanGe}
\usage{
\method{vcov}{fastHeckmanGe}(object, part = c("selection", "outcome", "dispersion", "correlation"))
}
\arguments{
\item{object}{fastHeckmanGe class object.}

\item{part}{A character vector with the model part names: selection, outcome, dispersion, correlation. By default it returns the complete vcov matrix, for all parts.}
}
\value{
The variance-covariance matrix of the coefficients of Generalized Heckman model.
}
\description{
Variance-covariance matrix of the Generalized Heckman Model
}
