% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovCL.fastHeckmanGe.R
\name{vcovCL.fastHeckmanGe}
\alias{vcovCL.fastHeckmanGe}
\title{vcovCL.fastHeckmanGe}
\usage{
vcovCL.fastHeckmanGe(
  x,
  cluster = NULL,
  type = NULL,
  sandwich = TRUE,
  fix = FALSE,
  ...
)
}
\description{
vcovCL.fastHeckmanGe
}
