% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.heckmanGE.R
\name{estfun.heckmanGE}
\alias{estfun.heckmanGE}
\title{Compute Estimating Functions for Generalized Heckman Model}
\usage{
estfun.heckmanGE(x, ...)
}
\arguments{
\item{x}{An object of class \code{heckmanGE}, which contains the fitted model, including model responses, model matrices, weights, and coefficient indexes.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A matrix of the same dimensions as the number of observations by the number of parameters in the model, where each element represents the gradient of the log-likelihood function with respect to the parameters.
}
\description{
This function calculates the estimating functions (gradient of the log-likelihood) for the Generalized Heckman model.
It is used for model diagnostics and for obtaining the estimating functions required for inference.
}
\details{
The function computes the gradient of the log-likelihood function for the Generalized Heckman model by extracting and manipulating various components of the model object.
It calculates the gradient based on the selection and outcome equations, as well as the dispersion and correlation components.

The function internally defines a helper function \code{gradlik_gen_i} to compute the gradient for each observation based on the model parameters.
}
