% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitheckmanGE.R
\name{fitheckmanGE}
\alias{fitheckmanGE}
\title{fitheckmanGE}
\usage{
fitheckmanGE(start, YS, XS, YO, XO, Msigma, Mrho, w)
}
\arguments{
\item{start}{A numeric vector with initial parameter values for the model.}

\item{YS}{A binary vector indicating the selection equation outcomes (1 for selected, 0 for not selected).}

\item{XS}{A matrix of predictors for the selection equation.}

\item{YO}{A vector of observed outcomes for the outcome equation.}

\item{XO}{A matrix of predictors for the outcome equation.}

\item{Msigma}{A matrix related to the dispersion parameter.}

\item{Mrho}{A matrix related to the correlation parameter.}

\item{w}{A numeric vector of weights for the observations.}
}
\value{
A list containing:
\describe{
\item{coefficients}{Estimated coefficients for the selection, outcome, dispersion, and correlation equations.}
\item{fitted.values}{Fitted values for the selection, outcome, dispersion, and correlation equations.}
\item{residuals}{Residuals for the selection and outcome equations.}
\item{loglik}{The log-likelihood of the fitted model.}
\item{vcov}{The variance-covariance matrix of the estimated parameters.}
\item{aic}{The Akaike Information Criterion (AIC) for the model.}
\item{bic}{The Bayesian Information Criterion (BIC) for the model.}
\item{optimization}{Details of the optimization process, including the convergence status.}
}
}
\description{
Optimized Function for Fitting the Generalized Heckman Model
}
\details{
This function estimates the parameters of a generalized Heckman selection model using a Newton-Raphson optimization algorithm. It supports the modeling of selection and outcome equations, along with associated dispersion and correlation structures.

The function optimizes the log-likelihood of a generalized Heckman selection model using a Newton-Raphson algorithm. The model allows for estimation of selection bias in the outcome equation, and incorporates additional parameters for dispersion and correlation.
}
