% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.heckmange.R
\name{predict.heckmanGE}
\alias{predict.heckmanGE}
\title{predict.heckmanGE
Predictions from the Generalized Heckman Model}
\usage{
\method{predict}{heckmanGE}(
  object,
  ...,
  part = c("selection", "outcome", "dispersion", "correlation"),
  newdata = NULL,
  type = c("link", "response"),
  cofint = FALSE,
  confidence_level = 0.95
)
}
\arguments{
\item{object}{An object of class \code{heckmanGE}. This object should be a fitted model from which predictions will be made.}

\item{\dots}{Additional arguments passed to specific methods. These are kept for compatibility with the generic \code{predict} function.}

\item{part}{A character vector specifying the model part for which to make predictions. Options include "selection", "outcome", "dispersion", and "correlation". Default is "outcome".}

\item{newdata}{Optionally, a data frame containing new data for making predictions. If omitted, the function uses the fitted linear predictors from the model object.}

\item{type}{The type of prediction required. Default is "link", which returns predictions on the scale of the linear predictors. If "response" is specified, predictions are on the scale of the response variable.}

\item{cofint}{A logical indicating whether to return confidence intervals for the predictions. Default is FALSE.}

\item{confidence_level}{A numeric value specifying the confidence level for the confidence intervals if \code{cofint} is TRUE. Default is 0.95.}
}
\value{
A vector or matrix of predictions. If \code{cofint} is TRUE, returns a matrix with predicted values and confidence intervals.
}
\description{
Generates predictions from a fitted \code{heckmanGE} model. Predictions can be made on the scale of the linear predictors or on the scale of the response variable. The function can also return confidence intervals for the predictions if requested.
}
