% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.heckmanGE.R
\name{print.heckmanGE}
\alias{print.heckmanGE}
\title{print.heckmanGE
Print of Generalized Heckman Model Results}
\usage{
\method{print}{heckmanGE}(x, ...)
}
\arguments{
\item{x}{An x of class \code{heckmanGE}. This x should be a fitted model whose results are to be printed.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Prints the estimates and statistics of the Generalized Heckman model to the console.
}
\description{
Prints a summary of the results from a fitted \code{heckmanGE} model, including estimates for different model components, log-likelihood, AIC, BIC, and other relevant statistics.
}
