% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.heckmange.R
\name{sandwich.heckmange}
\alias{sandwich.heckmange}
\title{sandwich.heckmange
Sandwich Estimator for Generalized Heckman Model}
\usage{
sandwich.heckmange(x, bread. = bread.heckmange, meat. = meat.heckmange, ...)
}
\arguments{
\item{x}{An object of class \code{heckmange} representing a fitted Generalized Heckman model.}

\item{bread.}{A function to compute the "bread" part of the sandwich estimator. Defaults to \code{bread.heckmange}.}

\item{meat.}{A function to compute the "meat" part of the sandwich estimator. Defaults to \code{meat.heckmange}.}

\item{...}{Additional arguments passed to \code{meat.}.}
}
\value{
A variance-covariance matrix for the \code{heckmange} model, computed using the sandwich estimator.
}
\description{
Computes the sandwich estimator of the variance-covariance matrix for the \code{heckmange} model.
}
